requirements.txt
streamlit
pandas
openpyxl
deep-translator
python-dotenv
